requirements.txt
imageio==2.34.1
matplotlib==3.8.4
nd2reader==3.3.0
nd2==0.10.1
numpy==2.0.1
opencv_python==4.10.0.84
PIMS==0.6.1
PyYAML==6.0.2
scipy==1.14.0
scikit-image==0.24.0